Renamed FindingsDashboard/utils to FindingsDashboard/findingsUtils for clarity.
